Add tests for models
Such tests make use of the refactored login commands, which means
that only the first tests needs to log in via the GUI, while others
benefit from the session to be stored from the cache.

In this test suite, I've also create a custom command to cleanup
all existing models on the beforeEach hook so that tests always
start in a fresh state.